requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
dict-toolbox==5.0.0
distro==1.9.0
Django==5.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
lazy-object-proxy==1.11.0
MarkupSafe==3.0.2
msgpack==1.1.0
nest-asyncio==1.6.0
openai==1.79.0
pop-config==12.0.4
pop-loop==1.1.0
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
rend==7.0.2
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
